requirements.txt
fastapi
pydantic
uvicorn
librosa
requests
grpcio
/code/riva_api-2.3.0-py3-none-any.whl